+++ app/src/main/java/com/example/roy/bossfit/Database/User.java
@@ -10,11 +10,11 @@ import android.arch.persistence.room.PrimaryKey;
 @Entity
 public class User {
     @PrimaryKey(autoGenerate = true)
-    private int id;
-    public int getId() {
+    private long id;
+    public long getId() {
         return id;
     }
-    public void setId(int id){
+    public void setId(long id){
         this.id=id;
     }
